Update Avalonia and to .NET 7

+++ src/MagicCollectionHelper/Core/Types/Domain.fs
@@ -12,7 +12,7 @@ module DomainTypes =
         cardGroupMinSize: uint
         cardGroupMaxSize: uint
         numBase: NumBase
-        missingPercent: float
+        missingPercent: decimal
         setWithFoils: bool
     }
